SC/requirements.txt
pandas==0.25.3
paho-mqtt
cryptography==2.6.1
asymcrypt
